requirements.txt
webrtcvad
noisereduce
pyaudio
numpy
openai
python-dotenv